requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cryptography==39.0.0
ecdsa==0.18.0
grpcio==1.51.1
grpcio-tools==1.51.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.0
pluggy==1.0.0
protobuf==4.21.12
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pytest==7.2.0
pytest-asyncio==0.20.3
python-jose==3.3.0
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
sanic==22.12.0
sanic-ext==22.12.0
sanic-routing==22.8.0
sanic-testing==22.12.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.4.0
ujson==5.7.0
uvloop==0.17.0
websockets==10.4
